requirements.txt
django
django-crispy-forms
bokeh
pandas
requests
ccxtnpm install ccxt